requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
slowapi==0.1.9
python-multipart==0.0.12
pillow==10.4.0
opencv-python-headless==4.10.0.84
doclayout-yolo==0.0.2
ollama==0.4.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
